==> 01-parse-original-data.ipynb <==
# `01-parse-original-data`

This notebook reads the original (if possible) source of the data, such as an Excel sheet provided as Supplementary Information to the cited reference. Since all the data sources use slightly different conventions there is a substantial amount of cleaning to get each into the same, simple format that the next notebooks then consume.

import pandas

pandas.options.display.max_columns=999

Define a `dict` so we can convert 3-letter amino acids to 1-letter codes

amino_acid_lookup = {'Cys': 'C', 'Asp': 'D', 'Ser': 'S', 'Gln': 'Q', 'Lys': 'K', 'Ile': 'I', 'Pro': 'P', 'Thr': 'T', 'Phe': 'F', 'Asn': 'N', 'Gly': 'G', 'His': 'H', 'Leu': 'L', 'Arg': 'R', 'Trp': 'W', 'Ala': 'A', 'Val':'V', 'Glu': 'E', 'Tyr': 'Y', 'Met': 'M', 'stop': '!'}

Load in the amino acid sequence of version 3 of H37Rv / NC_000962. Retrieved from mycobrowser.epfl.ch on 6 Jan 2023.

with open('data/reference/NC_000962.3_pncA_aminoacids.fasta','r') as INPUT:
    INPUT.readline()
    pncA_amino_acid_sequence = INPUT.readline()


# 1. Test/Train dataset

This is derived from three sources
(a). an in vitro mutagenesis study by Yadon et al. (Yadon2017)
(b). the NEJM resistance catalogue (NEJM2018)
(c). the WHO resistance catalogue (WHO2021)



### 1 (a). Yadon2017

YADON = pandas.read_excel('./data/source-data/41467_2017_721_MOESM4_ESM.xlsx')

# rename the columns to not contain spaces
YADON.rename(columns={  'Substitution':'MUTATION',
                        'Under-Represented?': 'UNDER_REPRESENTED',
                        'Catalogue': 'CATALOGUE',
                        'Miotto et. al.': 'MIOTTO2014',
                        'Walker et. al.': 'WALKER2015'}, inplace=True)

# make a Boolean IS_NONSYN columns
def define_nonsyn(row):
    IS_NONSYN = False
    IS_SYN = False
    CONTAINS_STOP = False
    if row['MUTATION'][-1]=='*':
        CONTAINS_STOP=True
    if row['MUTATION'][0]==row['MUTATION'][-1]:
        IS_SYN=True
    elif row['MUTATION'][0]!=row['MUTATION'][-1]:
        IS_NONSYN=True
    return pandas.Series([IS_NONSYN,IS_SYN,CONTAINS_STOP])

YADON[['IS_NONSYNONYMOUS','IS_SYNONYMOUS','CONTAINS_STOP']] = YADON.apply(define_nonsyn, axis=1)       
YADON['UNDER_REPRESENTED'] = YADON.UNDER_REPRESENTED=='Y'

# replace the NaNs with zeros and then convert to True/False                        
for i in ['in vitro Resistant', 'in vitro Susceptible', 'in vivo Resistant', 'in vivo Susceptible','MIOTTO2014','WALKER2015']:
    YADON[i] = YADON[i].fillna(0).astype('bool')                        

# combine the one-hot encoded columns
def assign_phenotypes(row, string):
    if row[string+' Resistant'] and not row[string+' Susceptible']:
        return('R')
    elif not row[string+' Resistant'] and row[string+' Susceptible']:
        return('S')
    else:
        return(None)

YADON['IN_VITRO_PHENOTYPE'] = YADON.apply(assign_phenotypes, args=('in vitro',), axis=1)
YADON['IN_VIVO_PHENOTYPE'] = YADON.apply(assign_phenotypes, args=('in vivo',), axis=1)

# drop the one-hot columns
YADON.drop(columns=['in vitro Resistant', 'in vitro Susceptible', 'in vivo Resistant', 'in vivo Susceptible'], inplace=True)

def interpet_catalogue(row):
    if row.CATALOGUE=='E':
        return('R')
    elif row.CATALOGUE=='D':
        return('S')

YADON['YADON_RESULT'] = YADON.apply(interpet_catalogue, axis=1)       

def replace_stop_codon(row):
    if "*" in row.MUTATION:
        return row.MUTATION.replace("*","!")
    else:
        return row.MUTATION

YADON['MUTATION'] = YADON.apply(replace_stop_codon, axis=1)

# reorder the dataset
YADON = YADON[['MUTATION', 'UNDER_REPRESENTED', 'IN_VITRO_PHENOTYPE', 'IN_VIVO_PHENOTYPE', 'YADON_RESULT', 'MIOTTO2014',
       'WALKER2015', 'IS_NONSYNONYMOUS','IS_SYNONYMOUS','CONTAINS_STOP']]

YADON = YADON[(~YADON.UNDER_REPRESENTED) & (YADON.YADON_RESULT.notna())]
YADON.set_index('MUTATION', inplace=True, verify_integrity=True)
YADON[['YADON_RESULT']].to_csv('data/catalogues/ds-yadon2017.csv', index=True)
YADON[['YADON_RESULT']][:3]

### 1 (b). NEJM2018

nejm2018 = pandas.read_csv('data/source-data/NC_000962.3_NEJM2018_v1.1_GARC1_RUS.csv')

nejm2018 = nejm2018[(nejm2018.DRUG=='PZA') & (nejm2018.PREDICTION.isin(['R','S'])) & (~nejm2018.MUTATION.str.contains('*', regex=False))]

def infer_type(row):
    mut = row.MUTATION
    cols = mut.split('@')
    IS_SNP = True
    IN_CDS = True
    if 'ins' in mut:
        IS_SNP=False
    elif 'del' in mut:
        IS_SNP=False
    elif '-' in mut:
        IN_CDS=False
    return pandas.Series([IS_SNP, IN_CDS, cols[0], cols[1]])        

nejm2018[['IS_SNP', 'IN_CDS', 'GENE', 'MUTATION']] = nejm2018.apply(infer_type, axis=1)    
nejm2018.set_index('MUTATION', inplace=True, verify_integrity=True)
nejm2018[['PREDICTION']].to_csv('data/catalogues/ds-nejm2018.csv', index=True)
nejm2018[['PREDICTION']][:3]

### 1 (c). WHO2021

who = pandas.read_csv('data/source-data/NC_000962.3_WHO-UCN-GTB-PCI-2021.7_v1.0_GARC1_RUS.csv')

who = who[(who.DRUG=='PZA') & (who.PREDICTION.isin(['R','S'])) & (~who.MUTATION.str.contains('*', regex=False))]

def infer_type(row):
    mut = row.MUTATION
    cols = mut.split('@')
    gene = cols[0]
    mutation = cols[1]
    IS_SNP = True
    IN_CDS = True
    if ('ins' in mut) or ('del' in mut):
        IS_SNP=False
        bits = mutation.split('_')
        mutation = bits[0]+"_indel"
    elif '-' in mut:
        IN_CDS=False
    return pandas.Series([IS_SNP, IN_CDS, gene, mutation])        

who[['IS_SNP', 'IN_CDS', 'GENE', 'MUTATION']] = who.apply(infer_type, axis=1)    

# By using a generic `X_indel` description there are now several different indels at a specific position, however they all confer resistance according to the catalogue so we can drop the duplicate rows
foo = who[~who.IS_SNP].MUTATION.value_counts()
print(who[who.MUTATION.isin(foo[foo>1].index)].PREDICTION.value_counts())

who.set_index('MUTATION', inplace=True, verify_integrity=False)
who = who.loc[~who.index.duplicated(keep='first')]
who.reset_index(inplace=True)
who.set_index('MUTATION', inplace=True, verify_integrity=True)
who[['PREDICTION']].to_csv('data/catalogues/ds-who2021.csv', index=True)
who[['PREDICTION']][:3]

# 2. Validation dataset

This is derived from three sets of clinical samples with DST data

### 2 (a). CRyPTIC2021

DST_CORE = pandas.read_csv('data/source-data/cryptic1-dst.csv')
DST_CORE.set_index('UNIQUEID', inplace=True)

LINEAGES_CORE = pandas.read_csv('data/source-data/cryptic1-lineages.csv')
LINEAGES_CORE.set_index('UNIQUEID', inplace=True)

MUTATIONS_CORE = pandas.read_csv('data/source-data/cryptic1-mutations.csv')
MUTATIONS_CORE.set_index('UNIQUEID', inplace=True)
MUTATIONS_CORE = MUTATIONS_CORE.join(LINEAGES_CORE, how='left')

print("There are %i samples in the CRyPTIC dataset with both genetics and a pncA phenotype" % len(DST_CORE))

If a sample has a mutation in *pncA*, it could have one, two or more than two mutations.

MUTATIONS_CORE.reset_index(inplace=True)
MUTATIONS_CORE_COUNTS=MUTATIONS_CORE[['UNIQUEID','SITEID']].groupby('UNIQUEID').count()
MUTATIONS_CORE.set_index('UNIQUEID',inplace=True)
MUTATIONS_CORE_COUNTS.rename(columns={'SITEID':'N_MUTATIONS'},inplace=True)
MUTATIONS_CORE_COUNTS
MUTATIONS_CORE_COUNTS.N_MUTATIONS.value_counts()

Some of these are likely to be lineage-associated. Let's check if any of the synoymous mutations are potential lineage-defining/associated mutations

mutation_counts = MUTATIONS_CORE.MUTATION.value_counts()

df = MUTATIONS_CORE[MUTATIONS_CORE.MUTATION.isin(mutation_counts[mutation_counts>=3].index) & MUTATIONS_CORE.IS_SYNONYMOUS]

pandas.crosstab(df.MUTATION, df.MYKROBE_LINEAGE_NAME_1)

Synonymous mutations S65S and A38A are entirely associated with different lineages and have n>10 so we will assume are lineage-associated and so can be ignored. This reduces the number of samples with >=1 mutation present.

MUTATIONS_CORE = MUTATIONS_CORE[~MUTATIONS_CORE.MUTATION.isin(['S65S','A38A'])]
MUTATIONS_CORE.reset_index(inplace=True)
MUTATIONS_CORE_COUNTS=MUTATIONS_CORE[['UNIQUEID','SITEID']].groupby('UNIQUEID').count()
MUTATIONS_CORE.set_index('UNIQUEID',inplace=True)
MUTATIONS_CORE_COUNTS.rename(columns={'SITEID':'N_MUTATIONS'},inplace=True)
MUTATIONS_CORE_COUNTS
MUTATIONS_CORE_COUNTS.N_MUTATIONS.value_counts()

Overall we end up with 3,611 samples with DST data and one or more mutations

samples_with_mutations=set(MUTATIONS_CORE.index)
DST_CORE.loc[DST_CORE.index.isin(samples_with_mutations), 'HAS_MUTATION']=True
DST_CORE.HAS_MUTATION.fillna(False,inplace=True)
DST_CORE.HAS_MUTATION.value_counts()

df = pandas.crosstab(DST_CORE.HAS_MUTATION, DST_CORE.PHENOTYPE, margins=True)
df

prop_R_mut = 100*df['R'][True] / df['All'][True]
prop_R_nomut = 100*df['R'][False] / df['All'][False]
print("%.1f %% of samples with a mutation are resistant, whilst only %.1f %% of samples without a mutation are resistant" % (prop_R_mut, prop_R_nomut))

Let's take a quick look at the most common mutations

MUTATIONS_CORE.MUTATION.value_counts()[:10]

Now we need to identify those samples which have one (and only one) mutation in *pncA*

DST_CORE=DST_CORE.join(MUTATIONS_CORE_COUNTS,how='left')
DST_CORE.N_MUTATIONS.fillna(0,inplace=True) 
DST_CORE['IS_SOLO']=DST_CORE.N_MUTATIONS==1
SOLOS=MUTATIONS_CORE.join(DST_CORE[DST_CORE.IS_SOLO][['PHENOTYPE', 'SOURCE']],how='inner')
SOLOS[:3]

Now we can summarise and write to disc

SOLO_MUTATIONS_CROSSTAB=pandas.crosstab([SOLOS.MUTATION,SOLOS.IS_SNP,SOLOS.IN_CDS],SOLOS.PHENOTYPE,margins=False,margins_name='TOTAL')
SOLO_MUTATIONS_CROSSTAB.to_csv('data/clinical-samples/ds-cryptic2021.csv')
SOLO_MUTATIONS_CROSSTAB

### 2 (b). Miotto2014 data

> Miotto, P., Cabibbe, A. M., Feuerriegel, S., Casali, N., Drobniewski, F., Rodionova, Y., Bakonyte, D., Stakenas, P., Pimkina, E., Augustynowicz-Kopeć, E., Degano, M., Ambrosi, A., Hoffner, S., Mansjö, M., Werngren, J., Rüsch-Gerdes, S., Niemann, S., & Cirillo, D. M. (2014). Mycobacterium tuberculosis pyrazinamide resistance determinants: a multicenter study. mBio, 5(5), e01819-14. https://doi.org/10.1128/mBio.01819-14

Load the raw Excel file from the supplement of this paper.

MIOTTO2014 = pandas.read_excel('./data/source-data/mbo005142032st1.xlsx', header=2)

def lookup_mutation(row):

    mutation = None
    IS_SNP = None
    IN_CDS = None
    
    # skipping rows with a comma skips samples with multiple mutations
    if row['pncA aa'] != 'WT' and ',' not in row['pncA nt'] and ',' not in row['pncA aa'] and row['pncA nt'] != 'del >200 nt':
        if row['pncA aa'][:3] not in amino_acid_lookup.keys():
            if 'promoter' in row['pncA aa']: # or 'shift' not in row['pncA aa']:
                IN_CDS = False
                if row['pncA nt'][:3]  in ['Ins', 'Del']: 
                    IS_SNP = False
                    mut = row['pncA nt']
                    pos = mut.split(' ')[0][3:]
                    mutation = (str(int(pos))+'_indel')
                else:
                    IS_SNP = True
                    mut = row['pncA nt']
                    assert mut[0] in ['a','t','c','g'], mut
                    assert mut[-1] in ['a','t','c','g'], mut
                    mutation = (row['pncA nt'])
        else:
            mut = row['pncA aa']
            IS_SNP = True
            IN_CDS = True
            if 'stop' in mut:
                if mut[:4] == 'stop':
                    mutation = (amino_acid_lookup[mut[:4]] + mut[4:-3] + amino_acid_lookup[mut[-3:]])
                else:
                    mutation = (amino_acid_lookup[mut[:3]] + mut[3:-4] + amino_acid_lookup[mut[-4:]])
            else:
                ref = amino_acid_lookup[mut[:3]]
                pos = mut[3:-3]
                alt = amino_acid_lookup[mut[-3:]]
                assert pncA_amino_acid_sequence[int(pos)-1] == ref, mut

                mutation = (ref + pos + alt)
    return pandas.Series([mutation, IS_SNP, IN_CDS])
    
MIOTTO2014[['MUTATION', 'IS_SNP', 'IN_CDS']] = MIOTTO2014.apply(lookup_mutation, axis=1)
MIOTTO2014.rename(columns = {'PZA': 'PHENOTYPE'}, inplace=True)
MIOTTO2014_MUTATIONS_SUMMARY=pandas.crosstab([MIOTTO2014.MUTATION,MIOTTO2014.IS_SNP,MIOTTO2014.IN_CDS],MIOTTO2014.PHENOTYPE,margins=False)
MIOTTO2014_MUTATIONS_SUMMARY

n_miotto = MIOTTO2014_MUTATIONS_SUMMARY.R.sum() + MIOTTO2014_MUTATIONS_SUMMARY.S.sum()
print("There are a total of " + str(n_miotto) + " samples in the Miotto2014 dataset")

MIOTTO2014_MUTATIONS_SUMMARY.reset_index(inplace=True)
n_miotto_nonsyn = MIOTTO2014_MUTATIONS_SUMMARY[(MIOTTO2014_MUTATIONS_SUMMARY.IS_SNP) & (MIOTTO2014_MUTATIONS_SUMMARY.IN_CDS)]
print("There are " + str(len(n_miotto_nonsyn)) + " non-synoymous mutations in " + str(n_miotto_nonsyn.R.sum() + n_miotto_nonsyn.S.sum()) + " samples")
MIOTTO2014_MUTATIONS_SUMMARY.set_index(['MUTATION', 'IS_SNP', 'IN_CDS'], inplace=True)
MIOTTO2014_MUTATIONS_SUMMARY.to_csv('data/clinical-samples/ds-miotto2014.csv',index=True)
MIOTTO2014_MUTATIONS_SUMMARY[:3]

### 2 (c). Whitfield2015

> Whitfield, M. G., Soeters, H. M., Warren, R. M., York, T., Sampson, S. L., Streicher, E. M., Van Helden, P. D., & Van Rie, A. (2015). A global perspective on pyrazinamide resistance: Systematic review and meta-analysis. PLoS ONE, 10(7), 1–16. https://doi.org/10.1371/journal.pone.0133869

Load in the Whitfield data -- this was copied from `zjm999094564so1.pdf`, missing out any promoter/indels, into an Excel sheet.

WHITFIELD2015 = pandas.read_excel('./data/source-data/zjm999094564so1.xlsx', header=1)

def parse_whitfield2015(row):

    mut = row.MUTATION.rstrip()
    IS_SNP = True
    IN_CDS = True

    if mut[:4]== 'stop':
        ref = amino_acid_lookup[mut[:4]]
        pos = mut[4:-3]
        alt = amino_acid_lookup[mut[-3:]]
    elif mut[-4:] == 'stop':
        ref = amino_acid_lookup[mut[:3]]
        pos = mut[3:-4]
        alt = amino_acid_lookup[mut[-4:]]
    else:
        ref = amino_acid_lookup[mut[:3]]
        pos = mut[3:-3]
        alt = amino_acid_lookup[mut[-3:]]

    mutation = ref + pos + alt
    if int(pos) <= 186:
        assert pncA_amino_acid_sequence[int(pos)-1] == ref, mut
    else:
        assert '!' == ref , mut

    
    return(pandas.Series([mutation, IS_SNP, IN_CDS]))

WHITFIELD2015[['MUTATION', 'IS_SNP', 'IN_CDS']] = WHITFIELD2015.apply(parse_whitfield2015, axis=1)
WHITFIELD2015 = WHITFIELD2015[WHITFIELD2015.MUTATION.notna()][['MUTATION', 'IS_SNP', 'IN_CDS', 'R', 'S']]
WHITFIELD2015.set_index(['MUTATION', 'IS_SNP', 'IN_CDS'], inplace=True, verify_integrity=True)
WHITFIELD2015.to_csv('data/clinical-samples/ds-whitfield2015.csv')
print("There are " + str(len(WHITFIELD2015)) + " non-synonymous mutations in " + str(WHITFIELD2015.R.sum() + WHITFIELD2015.S.sum()) + " samples in the Whitfield2015 dataset.")
WHITFIELD2015[:3]

## 3. Samples with MIC measurements

amino_acid_lookup = {'Cys': 'C', 'Asp': 'D', 'Ser': 'S', 'Gln': 'Q', 'Lys': 'K', 'Ile': 'I', 'Pro': 'P', 'Thr': 'T', 'Phe': 'F', 'Asn': 'N', 'Gly': 'G', 'His': 'H', 'Leu': 'L', 'Arg': 'R', 'Trp': 'W', 'Ala': 'A', 'Val':'V', 'Glu': 'E', 'Tyr': 'Y', 'Met': 'M', 'stop': '!'}

SAMPLE_MICS = pandas.read_csv('data/source-data/mic-measurements.csv')

def create_mutation(row):
    wt = amino_acid_lookup[row.WT.capitalize()]
    mt = amino_acid_lookup[row.MUT.capitalize()]
    return (wt + str(row.Residue) + mt)

SAMPLE_MICS['MUTATION'] = SAMPLE_MICS.apply(create_mutation, axis=1)

def parse_MIC(row):
    mic = row.MIC
    try:
        if (int(mic)) > 100:
            return "R"
        else:
            return "S"
    except:
        if mic=='>100':
            return "R"
        else:
            cols = mic.split('-')
            assert len(cols)==2
            if int(cols[1]) > 100:
                return "R"
            else:
                return "S"

SAMPLE_MICS['CONSISTENT_PHENOTYPE'] = SAMPLE_MICS.apply(parse_MIC, axis=1)  
SAMPLE_MICS.rename(columns={'Source':'SOURCE'}, inplace=True)

SAMPLE_MICS[:3]

print("There are %i isolates with one of %i mutations" % (len(SAMPLE_MICS), len(SAMPLE_MICS.MUTATION.unique())))

There are 10 mutations that are tested more than once; all of these are consistent except G97D and A102V

SAMPLE_MICS.set_index('MUTATION', inplace=True, verify_integrity=False)
SAMPLE_MICS[SAMPLE_MICS.index.duplicated(keep=False)][['MIC', 'CONSISTENT_PHENOTYPE']]

Therefore, remove G97D and A102V and keep just one measurement for the remaining duplicates

SAMPLE_MICS = SAMPLE_MICS[~SAMPLE_MICS.index.isin(['A102V', 'G97D'])]
SAMPLE_MICS = SAMPLE_MICS[~SAMPLE_MICS.index.duplicated(keep='last')]
SAMPLE_MICS.reset_index(inplace=True)
SAMPLE_MICS[:3]

print("This results in %i mutations of which %i are resistant and %i susceptible" % (len(SAMPLE_MICS), (SAMPLE_MICS.CONSISTENT_PHENOTYPE=='R').sum(), (SAMPLE_MICS.CONSISTENT_PHENOTYPE=='S').sum()) )

SAMPLE_MICS[['MUTATION', 'CONSISTENT_PHENOTYPE', 'MIC', 'SOURCE']].to_csv('data/ds-mic-phen.csv', index=False)
