==> 07-split-testtrain.ipynb <==
# `07-split-testtrain`

Note that since we have specified the random seed this is reproducible

import pandas, numpy, joblib

from sklearn.model_selection import train_test_split
from sklearn.preprocessing import RobustScaler
from sklearn.pipeline import Pipeline
from sklearn.base import BaseEstimator, TransformerMixin

import sbmlcore

random_state=42
mutations ={}
mapping_dict = {'S':0, 'R':1}

Let's read in the testtrain dataset so we can split it into training and test datasets but now we can drop the six features shown by the univariate logistical regression to be not much better than random

df = pandas.read_csv('data/ds-traintest-phen-features.csv')

df.drop(columns=['secondary_structure_codes',
 'd_MW',
 'phi',
 'd_volume',
 'd_Pi',
 'n_hbond_acceptors'],inplace=True)

mutations['tt'] = df['mutation']
features = df.iloc[:,3:]
labels = df['phenotype'].map({'S':0, 'R':1})
features[:3]

mutations_train, mutations_test, X_train, X_test, Y_train, Y_test = train_test_split(mutations['tt'], features, labels, test_size=0.3,
                                                    random_state=random_state)

class DataFrameSelector(BaseEstimator, TransformerMixin):
    def __init__(self, attribute_names):
        self.attribute_names = attribute_names
    def fit(self, X, y=None):
        return self
    def transform(self, X):
        return X[self.attribute_names].values

numerical_attribs = list(features)

headings = '\n'.join(i for i in numerical_attribs)

OUTPUT = open('data/ds-traintest-numericalattribs.txt', 'w')
OUTPUT.write(headings)
OUTPUT.close()

pipe = joblib.load('data/pipeline.pkl')

pipe = Pipeline([
    ('selector', DataFrameSelector(numerical_attribs)),
    ('scaling', RobustScaler())
    ])

X_train = pipe.fit_transform(X_train)    
Y_train = Y_train.squeeze().to_numpy()
Z_train = mutations_train.to_numpy()

# let's save the pipeline to preserve the scaling etc
joblib.dump(pipe, 'data/pipeline.pkl')

X_test = pipe.transform(X_test)
Y_test = Y_test.squeeze().to_numpy()   
Z_test = mutations_test.to_numpy()

with open('data/ds-train.npy', 'wb') as f:
    numpy.save(f, Y_train)
    numpy.save(f, X_train)
    numpy.save(f, Z_train)

with open('data/ds-test.npy', 'wb') as f:
    numpy.save(f, Y_test)
    numpy.save(f, X_test)    
    numpy.save(f, Z_test)    

df.loc[df.mutation.isin(mutations_test), 'set'] = 'test'
df.loc[df.mutation.isin(mutations_train), 'set'] = 'train'
df.to_csv('data/ds-traintest-phen-features-set.csv',index=False)

Let's check the % resistance is about the same between the train and test datasets

pandas.crosstab(df.set,df.phenotype,margins=True)

Now let's work out what the predictions using Suspect-PZA are for the training and test datasets for analysis later

suspectpza = pandas.read_csv('data/reference/ds-suspectpza-results.csv')
def create_mutation(row):
    return sbmlcore.amino_acid_3to1letter[row['Wild Type']] + str(row['Position']) + sbmlcore.amino_acid_3to1letter[row['Mutant']]

suspectpza['mutation'] = suspectpza.apply(create_mutation,axis=1)    
suspectpza.set_index('mutation', inplace=True)
suspectpza['prediction'] = suspectpza[' Prediction'].str[0].map(mapping_dict)
suspectpza.drop(columns=['Chain','Wild Type','Position','Mutant','Distance to ligand',' Prediction'],inplace=True)
suspectpza

Using the (fitted) pipeline we can repeat the process for the two types of `validation` dataset, depending whether just clinical samples or aggregated mutations

for i in ['samples','mutations']:
    df = pandas.read_csv('data/ds-validation-' + i + '-phen-features.csv')

    df.drop(columns=['secondary_structure_codes',
        'd_MW',
        'phi',
        'd_volume',
        'd_Pi',
        'n_hbond_acceptors'],inplace=True)

    mutations['v'] = df['mutation']
    features = df.iloc[:,3:]
    labels = df['phenotype'].map({'S':0, 'R':1, 'U':2})

    X_validate = pipe.transform(features)
    Y_validate = labels.squeeze().to_numpy()
    Z_validate = mutations['v'].to_numpy()

    with open('data/ds-validation-' + i + '.npy', 'wb') as f:
        numpy.save(f, Y_validate)
        numpy.save(f, X_validate)
        numpy.save(f, Z_validate)

..and again label the mutations present in the `validation` dataset

df.set_index('mutation', inplace=True)
suspectpza = suspectpza.join(df[['phenotype']], how='outer')
suspectpza['validation-mutations'] = suspectpza.phenotype.map(mapping_dict)
suspectpza.drop(columns=['phenotype'], inplace=True)
suspectpza

df = pandas.read_csv('data/ds-mic-phen-features.csv')

df.drop(columns=['secondary_structure_codes',
        'd_MW',
        'phi',
        'd_volume',
        'd_Pi',
        'n_hbond_acceptors'],inplace=True)


mutations['mic'] = df['mutation']
features = df.iloc[:,3:]
labels = df['phenotype'].map({'S':0, 'R':1})

X_mic = pipe.transform(features)
Y_mic = labels.squeeze().to_numpy()
Z_mic = mutations['mic'].to_numpy()

with open('data/ds-mic.npy', 'wb') as f:
    numpy.save(f, Y_mic)
    numpy.save(f, X_mic)
    numpy.save(f, Z_mic)

df.set_index('mutation', inplace=True)
suspectpza = suspectpza.join(df[['phenotype']], how='outer')
suspectpza['mic'] = suspectpza.phenotype.map(mapping_dict)
suspectpza.drop(columns=['phenotype'], inplace=True)
suspectpza[:3]

assert len(Y_mic) == suspectpza.mic.notna().sum(), suspectpza.mic.notna().sum()
assert len(Y_validate) == suspectpza['validation-mutations'].notna().sum()

suspectpza

for i in ['validation-mutations', 'mic']:
    Y=suspectpza[suspectpza[i].notna()][i].squeeze().to_numpy()
    Z=suspectpza[suspectpza[i].notna()].prediction.squeeze().to_numpy()
    ZZ=suspectpza[suspectpza[i].notna()].index.to_numpy()
    with open('data/suspectpza-'+i+'.npy', 'wb') as f:
        numpy.save(f, Y)
        numpy.save(f, Z)
        numpy.save(f, ZZ)

df = pandas.read_csv('data/ds-validation-samples-phen.csv')
df.rename(columns={'MUTATION': 'mutation', 'PHENOTYPE': 'phenotype'}, inplace=True)
df.set_index('mutation',inplace=True)
df[:3]

suspectpza = suspectpza.join(df[['phenotype']], how='outer')
suspectpza.prediction.value_counts(dropna=False)

suspectpza['validation-samples'] = suspectpza.phenotype.map(mapping_dict)
suspectpza[:3]

for i in ['validation-samples']:
    Y=suspectpza[suspectpza[i].notna()][i].squeeze().to_numpy()
    Z=suspectpza[suspectpza[i].notna()].prediction.squeeze().to_numpy()
    ZZ=suspectpza[suspectpza[i].notna()].index.to_numpy()
    with open('data/suspectpza-'+i+'.npy', 'wb') as f:
        numpy.save(f, Y)
        numpy.save(f, Z)
        numpy.save(f, ZZ)

suspectpza = suspectpza[~(suspectpza['validation-mutations'].isna() & suspectpza.mic.isna())]
suspectpza.to_csv('data/reference/ds-suspectpza-results-sets.csv')
