==> 12-create-figure-5.ipynb <==
# `12-create-figure-5`

import pandas, numpy, pathlib

import matplotlib.pyplot as plt
from matplotlib.patches import Rectangle

from skops.io import load

from misc import construct_line

pathlib.Path('pdf').mkdir(exist_ok=True)
pathlib.Path('pdf/figure-5').mkdir(exist_ok=True)

chosen_model = 'XB'

best_model = {}

for model in ['LR', 'NN', 'XB']:
    best_model[model] = load('models/'+model.lower()+'.skops', trusted=True)

X={}
Y={}
Z={}

for i in ['validation-samples', 'validation-mutations', 'mic', 'train', 'test']:
    X[i]={}
    Y[i]={}
    Z[i]={}
    with open('data/ds-'+i+'.npy', 'rb') as f:
        Y[i]['input'] = numpy.load(f)
        X[i]['input'] = numpy.load(f)
        Z[i]['input'] = numpy.load(f, allow_pickle=True)

# load in the results for SuspectPZA
suspectpza={}
for i in ['validation-samples', 'validation-mutations','mic']:
    suspectpza[i]={}
    with open('data/suspectpza-'+i+'.npy', 'rb') as f:
        suspectpza[i]['input'] = numpy.load(f)
        suspectpza[i]['predicted'] = numpy.load(f)
        suspectpza[i]['muts'] = numpy.load(f, allow_pickle=True)
line = []

def validate_model(line, best_model, model_name, dataset, X, Y):
    
    Y[dataset]['predicted'] = best_model.predict(X[dataset]['input'])
    Y[dataset]['scores'] = best_model.predict_proba(X[dataset]['input'])[:,1]

    row = construct_line(model_name , dataset, None, Y[dataset], None)

    n_muts_unique = (numpy.unique(Z[dataset]['input'])).shape[0]

    row.append(X[dataset]['input'].shape[0])
    row.append(n_muts_unique)
    line.append(row)

    return(line)

First, let's work out which of the mutations in the Validation dataset have inconsistent mutations

validation_mutations = pandas.read_csv('data/ds-validation-mutations-full.csv')

mutations_with_consistent_phenotypes= set(validation_mutations[validation_mutations.CONSISTENT_PHENOTYPE.isin(['R', 'S'])].MUTATION.unique())

mutations_with_inconsistent_phenotypes = set(validation_mutations[validation_mutations.CONSISTENT_PHENOTYPE=='U'].MUTATION.unique())

len(mutations_with_inconsistent_phenotypes), len(mutations_with_consistent_phenotypes)

mask = numpy.isin(Z['validation-samples']['input'], numpy.array(list(mutations_with_inconsistent_phenotypes)))

X['vs-incons']={}
Y['vs-incons']={}
Z['vs-incons']={}

X['vs-incons']['input']=X['validation-samples']['input'][mask]
Y['vs-incons']['input']=Y['validation-samples']['input'][mask]
Z['vs-incons']['input']=Z['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-samples']['input'], numpy.array(list(mutations_with_consistent_phenotypes)))

X['vs-consis']={}
Y['vs-consis']={}
Z['vs-consis']={}

X['vs-consis']['input']=X['validation-samples']['input'][mask]
Y['vs-consis']['input']=Y['validation-samples']['input'][mask]
Z['vs-consis']['input']=Z['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-mutations']['input'], numpy.array(list(mutations_with_consistent_phenotypes)))

X['vm-consis']={}
Y['vm-consis']={}
Z['vm-consis']={}

X['vm-consis']['input']=X['validation-mutations']['input'][mask]
Y['vm-consis']['input']=Y['validation-mutations']['input'][mask]
Z['vm-consis']['input']=Z['validation-mutations']['input'][mask]

# mask = numpy.isin(Z['validation-mutations']['input'], numpy.array(list(mutations_with_consistent_phenotypes)))

# X['vm-incons']={}
# Y['vm-incons']={}
# Z['vm-incons']={}

# X['vm-incons']['input']=X['validation-mutations']['input'][mask]
# Y['vm-incons']['input']=Y['validation-mutations']['input'][mask]
# Z['vm-incons']['input']=Z['validation-mutations']['input'][mask]


for dataset in ['validation-samples', 'validation-mutations',  'mic', 'vs-incons', 'vs-consis', 'vm-consis']: #, 'vm-incons']: 
    line = validate_model(line, best_model[chosen_model], chosen_model, dataset, X, Y)

total = validation_mutations.TOTAL.sum()

print("There are %i samples in the validation dataset" % (total))

total = validation_mutations[validation_mutations.CONSISTENT_PHENOTYPE=='U'].TOTAL.sum()

print("Of these, %i have a mutation which has an inconsistent phenotype" % (total))

total_lt4 = validation_mutations[(validation_mutations.CONSISTENT_PHENOTYPE=='U') & (validation_mutations.TOTAL<4)].TOTAL.sum()
total_ge4 = validation_mutations[(validation_mutations.CONSISTENT_PHENOTYPE=='U') & (validation_mutations.TOTAL>=4)].TOTAL.sum()
print("And %i of these are samples were there are fewer than 4 samples and %i are samples with five or more samples" % (total_lt4, total_ge4))

mutations_in_train  = set(Z['train']['input'])  #.union(set(Z['test']['input']))

mutations_with_consistent_phenotypes_in_train = mutations_with_consistent_phenotypes.intersection(mutations_in_train)
mutations_with_consistent_phenotypes_notin_train = mutations_with_consistent_phenotypes - mutations_in_train

mutations_with_inconsistent_phenotypes_in_train = mutations_with_inconsistent_phenotypes.intersection(mutations_in_train)
mutations_with_inconsistent_phenotypes_notin_train = mutations_with_inconsistent_phenotypes - mutations_in_train 

len(mutations_with_consistent_phenotypes_in_train), len(mutations_with_consistent_phenotypes_notin_train), len(mutations_with_inconsistent_phenotypes_in_train), len(mutations_with_inconsistent_phenotypes_notin_train)

for i in ['vs-consis-train', 'vs-consis-notrain', 'vs-incons-train', 'vs-incons-notrain', 'vm-consis-train', 'vm-consis-notrain', 'vm-incons-train', 'vm-incons-notrain']:
    X[i] = {}
    Y[i] = {}
    Z[i] = {}

mask = numpy.isin(Z['validation-samples']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_in_train)))
X['vs-consis-train']['input']=X['validation-samples']['input'][mask]
Y['vs-consis-train']['input']=Y['validation-samples']['input'][mask]
Z['vs-consis-train']['input']=Z['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-samples']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_notin_train)))
X['vs-consis-notrain']['input']=X['validation-samples']['input'][mask]
Y['vs-consis-notrain']['input']=Y['validation-samples']['input'][mask]
Z['vs-consis-notrain']['input']=Z['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-samples']['input'],  numpy.array(list(mutations_with_inconsistent_phenotypes_in_train)))
X['vs-incons-train']['input']=X['validation-samples']['input'][mask]
Z['vs-incons-train']['input']=Z['validation-samples']['input'][mask]
Y['vs-incons-train']['input']=Y['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-samples']['input'],  numpy.array(list(mutations_with_inconsistent_phenotypes_notin_train)))
X['vs-incons-notrain']['input']=X['validation-samples']['input'][mask]
Z['vs-incons-notrain']['input']=Z['validation-samples']['input'][mask]
Y['vs-incons-notrain']['input']=Y['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-mutations']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_in_train)))
X['vm-consis-train']['input']=X['validation-mutations']['input'][mask]
Y['vm-consis-train']['input']=Y['validation-mutations']['input'][mask]
Z['vm-consis-train']['input']=Z['validation-mutations']['input'][mask]


mask = numpy.isin(Z['validation-mutations']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_in_train)))
X['vm-consis-train']['input']=X['validation-mutations']['input'][mask]
Y['vm-consis-train']['input']=Y['validation-mutations']['input'][mask]
Z['vm-consis-train']['input']=Z['validation-mutations']['input'][mask]

mask = numpy.isin(Z['validation-mutations']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_notin_train)))
X['vm-consis-notrain']['input']=X['validation-mutations']['input'][mask]
Y['vm-consis-notrain']['input']=Y['validation-mutations']['input'][mask]
Z['vm-consis-notrain']['input']=Z['validation-mutations']['input'][mask]


# mask = numpy.isin(Z['validation-mutations']['input'],  numpy.array(list(mutations_with_inconsistent_phenotypes_in_train)))
# X['vm-incons-train']['input']=X['validation-mutations']['input'][mask]
# Z['vm-incons-train']['input']=Z['validation-mutations']['input'][mask]
# Y['vm-incons-train']['input']=Y['validation-mutations']['input'][mask]

# mask = numpy.isin(Z['validation-mutations']['input'],  numpy.array(list(mutations_with_inconsistent_phenotypes_notin_train)))
# X['vm-incons-notrain']['input']=X['validation-mutations']['input'][mask]
# Z['vm-incons-notrain']['input']=Z['validation-mutations']['input'][mask]
# Y['vm-incons-notrain']['input']=Y['validation-mutations']['input'][mask]


for dataset in ['vs-consis-train', 'vs-consis-notrain', 'vs-incons-train', 'vs-incons-notrain', 'vm-consis-train', 'vm-consis-notrain']: # 'vm-incons-train', 'vm-incons-notrain']:
    print(dataset)
    line = validate_model(line, best_model[chosen_model], chosen_model, dataset, X, Y)

mutations_with_consistent_phenotypes_notin_train_large = set(list(validation_mutations[(validation_mutations.MUTATION.isin(mutations_with_consistent_phenotypes_notin_train)) & (validation_mutations.TOTAL>=4)].MUTATION.unique()))
mutations_with_consistent_phenotypes_notin_train_small = set(list(validation_mutations[(validation_mutations.MUTATION.isin(mutations_with_consistent_phenotypes_notin_train)) & (validation_mutations.TOTAL<4)].MUTATION.unique()))

for i in ['vs-consis-notrain-large', 'vs-consis-notrain-small']:
    X[i] = {}
    Y[i] = {}
    Z[i] = {}

mask = numpy.isin(Z['validation-samples']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_notin_train_large)))
X['vs-consis-notrain-large']['input']=X['validation-samples']['input'][mask]
Z['vs-consis-notrain-large']['input']=Z['validation-samples']['input'][mask]
Y['vs-consis-notrain-large']['input']=Y['validation-samples']['input'][mask]

mask = numpy.isin(Z['validation-samples']['input'],  numpy.array(list(mutations_with_consistent_phenotypes_notin_train_small)))
X['vs-consis-notrain-small']['input']=X['validation-samples']['input'][mask]
Z['vs-consis-notrain-small']['input']=Z['validation-samples']['input'][mask]
Y['vs-consis-notrain-small']['input']=Y['validation-samples']['input'][mask]

for dataset in ['vs-consis-notrain-large', 'vs-consis-notrain-small']:
    print(dataset)
    line = validate_model(line, best_model[chosen_model], chosen_model, dataset, X, Y)

'resid '+ ' '.join([i[1:-1] for i in mutations_with_consistent_phenotypes_notin_train_large])

'resid '+ ' '.join([i[1:-1] for i in mutations_with_consistent_phenotypes_notin_train_small])

test_results = pandas.DataFrame(line, columns=['model', 'dataset', 'sensitivity_mean', 'sensitivity_std', 'specificity_mean', 'specificity_std' ,'roc_auc_mean', 'roc_auc_std','TN','FP','FN','TP', 'model_parameters', 'n_samples', 'n_mutations'])
test_results

for dataset in list(test_results.dataset):
    print(dataset)
    colour='#888888'
    fig = plt.figure(figsize=(1.8, 3.5))
    axes = plt.gca()
    axes.spines['top'].set_visible(False)
    axes.spines['right'].set_visible(False)
    axes.spines['left'].set_visible(False)
    axes.get_yaxis().set_visible(False)
    x=[0,1]
    y=test_results[(test_results.dataset==dataset)][['sensitivity_mean','specificity_mean']].values[0]
    axes.set_ylim([0,100])
    axes.bar(x,y, label=y, edgecolor=colour, color='None',linewidth=2)
    axes.set_xticks([0,1])
    axes.set_xticklabels(['sensitivity', 'specificity'])
    for (i,j) in zip(x,y):
        axes.text(i,j+2,'%.1f' % j,ha='center',color=colour)
    fig.savefig('pdf/figure-5/fig-5-'+str(dataset)+'.pdf',bbox_inches='tight',transparent=True)
    plt.close()

for idx,row in test_results.iterrows():

    fig = plt.figure(figsize=(1.5, 1.5))
    axes = plt.gca()

    axes.add_patch(Rectangle((0,0),1,1,fc='#e41a1c',alpha=0.7))
    axes.add_patch(Rectangle((0,1),1,1,fc='#4daf4a',alpha=0.7))
    axes.add_patch(Rectangle((1,1),1,1,fc='#fc9272',alpha=0.7))
    axes.add_patch(Rectangle((1,0),1,1,fc='#4daf4a',alpha=0.7))

    axes.set_xlim([0,2])
    axes.set_ylim([0,2])

    axes.set_xticks([0.5,1.5],labels=['R','S'])
    axes.set_yticks([0.5,1.5],labels=['S','R'])

    axes.text(0.5,0.5,row['FN'],ha='center',va='center')
    axes.text(1.5,0.5,row['TN'],ha='center',va='center')
    axes.text(0.5,1.5,row['TP'],ha='center',va='center')
    axes.text(1.5,1.5,row['FP'],ha='center',va='center')

    fig.savefig('pdf/figure-5/truthtable-'+row['dataset']+'-'+row['model']+'.pdf',bbox_inches='tight',transparent=True)
    plt.close()

total_lt4 = validation_mutations[(validation_mutations.MUTATION.isin(mutations_with_consistent_phenotypes_notin_train)) & (validation_mutations.TOTAL>=4)].TOTAL.sum()
len_lt4 = len(validation_mutations[(validation_mutations.MUTATION.isin(mutations_with_consistent_phenotypes_notin_train)) & (validation_mutations.TOTAL>=4)])
total_lt4, len_lt4

validation_mutations
